change FPGA kernel pointer type close #781

+++ src/operators/kernel/fpga/conv_bn_kernel.cpp
@@ -56,7 +56,7 @@ bool ConvBNKernel<FPGA, float>::Init(FusionConvBNParam *param) {
   param->SetNewScale(new_scale);
   param->SetNewBias(new_bias);
   fpga::quantify_filter(filter);
-  auto filter_ptr = filter->data<float>();
+  auto filter_ptr = filter->data<int8_t>();
 
   fpga::ConvArgs convArgs;
   convArgs.relu_enabled = relu_enabled;
